feat: 간단진단 구현 및 포스트맨 테스트 완료

+++ src/main/java/com/example/SeaTea/domain/diagnosis/converter/DiagnosisQuickConverter.java
@@ -49,7 +49,7 @@ public class DiagnosisQuickConverter {
     public static DiagnosisQuickResponseDTO toResponseDTO(
             TastingNoteTypeCode resultType,
             List<QuickKeyword> keywords,
-            Map<TastingNoteTypeCode, Double> scores
+            Map<TastingNoteTypeCode, Integer> scores
     ) {
         return DiagnosisQuickResponseDTO.builder()
                 .resultType(resultType)
